enable moltiple projetcs and testscripts for stwcli builder

+++ src/main/java/org/jenkins/plugins/silktestsuite/common/Utils.java
@@ -24,11 +24,13 @@ public class Utils {
         }
 
         private void deleteRecursive(File file) {
-          if (file.isFile() && (file.lastModified() < startTime.getTimeInMillis()))
-            file.delete();
-          else {
-            for (File f : file.listFiles()) {
-              deleteRecursive(f);
+          if (file != null) {
+            if (file.isFile() && (file.lastModified() < startTime.getTimeInMillis()))
+              file.delete();
+            else {
+              for (File f : file.listFiles()) {
+                deleteRecursive(f);
+              }
             }
           }
         }
